requirements.txt
pandas==1.3.5
numpy==1.21.5
tensorflow==2.8.0
keras==2.8.0
scikit-learn==1.0.2
matplotlib==3.5.1
seaborn==0.11.2
cufflinks==0.17.3
statsmodels==0.13.2
ipykernel==6.9.1
ipywidgets==7.6.5
joblib==1.1.0
PyYAML==6.0
scipy==1.7.3
ipython==7.31.1
chart-studio==1.1.0
